Rename classes: Service, AccountService.java. Changed the constructor of AccountRepository.java
