+++ src/main/java/com/sbrf/reboot/service/AccountService.java
@@ -1,7 +1,21 @@
 package com.sbrf.reboot.service;
 
 import com.sbrf.reboot.Account;
+import com.sbrf.reboot.repository.AccountRepository;
 
-public interface AccountService {
-    boolean isAccountExist(long l, Account account);
+public class AccountService implements Service {
+
+    private AccountRepository accountRepository;
+
+
+    public AccountService(AccountRepository accountRepository) {
+        this.accountRepository = accountRepository;
+    }
+
+
+    @Override
+    public boolean isAccountExist(long clientID, Account account) {
+        return accountRepository.getAllAccountsByClientId(clientID).contains(account);
+    }
+
 }
