+++ src/main/java/com/sbrf/reboot/service/Service.java
@@ -1,0 +1,7 @@
+package com.sbrf.reboot.service;
+
+import com.sbrf.reboot.Account;
+
+public interface Service {
+    boolean isAccountExist(long l, Account account);
+}
